requirements.txt
adamp==0.3.0
apex==0.1
async-timeout==3.0.1
blessed==1.19.1
blessings==1.7
blinker==1.4
brotlipy==0.7.0
certifi==2021.10.8
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.30
diffdist==0.1
dm-tree==0.1.6
docker-pycreds==0.4.0
easydict==1.9
Flask==2.1.2
fonttools==4.28.3
gast==0.5.3
gitdb==4.0.9
gspread==5.4.0
h5py==3.7.0
httplib2==0.20.4
ImageHash==4.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.1.1
matplotlib==3.5.0
mkl-fft==1.3.1
mkl-service==2.4.0
mnist==0.2.2
munch==2.5.0
nltk==3.7
nose==1.3.7
nvidia-ml-py==11.495.46
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
opencv-python==4.5.4.58
packaging==21.3
pandas==1.3.4
pathtools==0.1.2
Pillow==8.4.0
progressbar==2.5
promise==2.3
protobuf==3.19.1
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycocotools==2.0.4
pycodestyle==2.8.0
pynvml==11.4.1
pyparsing==3.0.6
pytz==2021.3
PyWavelets==1.4.0
PyYAML==6.0
pyzmq==19.0.2
regex==2022.6.2
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
sentencepiece==0.1.96
sentry-sdk==1.9.5
setproctitle==1.3.2
setuptools-scm==6.3.2
shortuuid==1.0.9
sklearn==0.0
smmap==5.0.0
tensorboard-logger==0.1.0
tensorflow-probability==0.16.0
termcolor==1.1.0
threadpoolctl==3.0.0
timm==0.3.2
tomli==1.2.2
torch==1.10.0+cu113
torch-tb-profiler==0.3.1
torchaudio==0.10.0+cu113
torchlars==0.1.2
torchtext==0.11.0
torchvision==0.11.1+cu113
tqdm==4.62.3
urllib3==1.26.11
wandb==0.13.1
